config.js set to new dev-backend-url, Readme enhanced with information about configuring BACKEND_URL, Components directory moved src root
