+++ README.md
@@ -42,6 +42,9 @@ Instead, it will copy all the configuration files and the transitive dependencie
 
 You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
 
+## Docker
+To configure backend URL you can specify it by setting docker ENV `BACKEND_URL`. On runtime, docker-starter.sh would extract iot an write to config.js in public directory
+
 ## Learn More
 
 You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
