+++ public/config.js
@@ -1,2 +1,2 @@
 // this is dummy file for dev purpose
-window.BACKEND_URL = 'https://dev-lockdown-app-backend-hesbetn2tq-ew.a.run.app/api'
+window.BACKEND_URL='https://dev-safeticket-api.herokuapp.com/api'
