+++ public/index.html
@@ -40,9 +40,7 @@
     <meta name="msapplication-TileColor" content="#ffffff">
     <meta name="msapplication-TileImage" content="/ms-icon-144x144.png">
     <meta name="theme-color" content="#ffffff">
-
     <script src="%PUBLIC_URL%/config.js"></script>
-
     <title>SafeTicket</title>
   </head>
   <body>
